requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.6
jsonschema==4.17.1
drf-spectacular==0.26.3
